third_party/WebKit/LayoutTests/fast/dom/Document/document-elementFromPoint-on-option-element.html

<!DOCTYPE html>
<html>
<script src="../../../resources/js-test.js"></script>
<body>
<div id="container">
<select size="2">
  <option value="1">one</select>
  <option value="2">two</select>
  <option value="3">three</select>
</select>
</div>
<div id="console"></div>
</body>
</html>
<script>
description('Document.elementPoint() on option element');
var select = document.querySelector('select');
var result = document.elementFromPoint(select.offsetLeft + 5, select.offsetTop + 5);
shouldBeEqualToString('result.tagName', 'OPTION');
if (window.testRunner)
  document.getElementById('container').outerHTML = '';
</script>
