third_party/blink/web_tests/fast/dom/Document/document-elementFromPoint-on-option-element-expected.txt

Document.elementPoint() on option element

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result.tagName is "OPTION"
PASS successfullyParsed is true

TEST COMPLETE
